CountdownView tap to change to absolute date
